% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-lib-loads.R
\name{core_packages}
\alias{core_packages}
\title{Functions to Install all Core Libraries}
\usage{
core_packages()
}
\value{
A character vector
}
\description{
Lists the core packages necessary to run all potential modeling
algorithms.
}
\details{
Lists the core packages necessary to run all potential modeling
algorithms.
}
\examples{
core_packages()

}
\seealso{
Other Utility:
\code{\link[=check_duplicate_rows]{check_duplicate_rows()}},
\code{\link[=create_splits]{create_splits()}},
\code{\link[=create_workflow_set]{create_workflow_set()}},
\code{\link[=fast_classification_parsnip_spec_tbl]{fast_classification_parsnip_spec_tbl()}},
\code{\link[=fast_regression_parsnip_spec_tbl]{fast_regression_parsnip_spec_tbl()}},
\code{\link[=full_internal_make_wflw]{full_internal_make_wflw()}},
\code{\link[=install_deps]{install_deps()}},
\code{\link[=load_deps]{load_deps()}},
\code{\link[=match_args]{match_args()}},
\code{\link[=quantile_normalize]{quantile_normalize()}}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Utility}
