% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals-make-spec-tbl.R
\name{internal_make_spec_tbl}
\alias{internal_make_spec_tbl}
\title{Internals Make a Model Spec tibble}
\usage{
internal_make_spec_tbl(.model_tbl)
}
\arguments{
\item{.model_tbl}{This is the data that should be coming from inside of the
regression/classification to parsnip spec functions.}
}
\value{
A model spec tbl.
}
\description{
Make a Model Spec tibble.
}
\details{
Make a Model Spec tibble.
}
\examples{
make_regression_base_tbl() |>
  internal_make_spec_tbl()

make_classification_base_tbl() |>
  internal_make_spec_tbl()

}
\seealso{
Other Internals:
\code{\link[=internal_make_fitted_wflw]{internal_make_fitted_wflw()}},
\code{\link[=internal_make_wflw]{internal_make_wflw()}},
\code{\link[=internal_make_wflw_gee_lin_reg]{internal_make_wflw_gee_lin_reg()}},
\code{\link[=internal_make_wflw_predictions]{internal_make_wflw_predictions()}},
\code{\link{internal_model_builders_classification}},
\code{\link{internal_model_builders_regression}},
\code{\link[=internal_set_args_to_tune]{internal_set_args_to_tune()}},
\code{\link[=make_classification_base_tbl]{make_classification_base_tbl()}},
\code{\link[=make_regression_base_tbl]{make_regression_base_tbl()}}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Internals}
