% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals-model-builders-classification.R
\name{internal_model_builders_classification}
\alias{internal_model_builders_classification}
\alias{earth_classification_bag_mars}
\alias{earth_classification_mars}
\alias{earth_classification_discrim_flexible}
\alias{dbarts_classification_bart}
\alias{mass_classification_discrim_linear}
\alias{mda_classification_discrim_linear}
\alias{sda_classification_discrim_linear}
\alias{sparsediscrim_classification_discrim_linear}
\alias{mass_classification_discrim_quad}
\alias{sparsediscrim_classification_discrim_quad}
\alias{klar_classification_discrim_regularized}
\alias{mgcv_classification_gen_additive_mod}
\alias{brulee_classification_logistic_reg}
\alias{gee_classification_logistic_reg}
\alias{glm_classification_logistic_reg}
\alias{glmer_classification_logistic_reg}
\alias{glmnet_classification_logistic_reg}
\alias{liblinear_classification_logistic_reg}
\alias{brulee_classification_mlp}
\alias{nnet_classification_mlp}
\alias{brulee_classification_multinom_reg}
\alias{glmnet_classification_multinom_reg}
\alias{nnet_classification_multinom_reg}
\alias{klar_classification_naive_bayes}
\alias{kknn_classification_nearest_neighbor}
\alias{xrf_classification_rule_fit}
\alias{kernlab_classification_svm_linear}
\alias{liblinear_classification_svm_linear}
\alias{kernlab_classification_svm_poly}
\alias{kernlab_classification_svm_rbf}
\alias{liquidsvm_classification_svm_rbf}
\title{Internal Model Builders for Classification}
\usage{
earth_classification_bag_mars()

earth_classification_mars()

earth_classification_discrim_flexible()

dbarts_classification_bart()

mass_classification_discrim_linear()

mda_classification_discrim_linear()

sda_classification_discrim_linear()

sparsediscrim_classification_discrim_linear()

mass_classification_discrim_quad()

sparsediscrim_classification_discrim_quad()

klar_classification_discrim_regularized()

mgcv_classification_gen_additive_mod()

brulee_classification_logistic_reg()

gee_classification_logistic_reg()

glm_classification_logistic_reg()

glmer_classification_logistic_reg()

glmnet_classification_logistic_reg()

liblinear_classification_logistic_reg()

brulee_classification_mlp()

nnet_classification_mlp()

brulee_classification_multinom_reg()

glmnet_classification_multinom_reg()

nnet_classification_multinom_reg()

klar_classification_naive_bayes()

kknn_classification_nearest_neighbor()

xrf_classification_rule_fit()

kernlab_classification_svm_linear()

liblinear_classification_svm_linear()

kernlab_classification_svm_poly()

kernlab_classification_svm_rbf()

liquidsvm_classification_svm_rbf()
}
\value{
A parsnip model specification object
}
\description{
Internal functions that build pre-configured parsnip model specifications
for classification models. Each function returns a parsnip model specification
with the mode and engine pre-set.
}
\details{
These functions are used internally by \code{internal_make_spec_tbl()} to create
model specifications. Each function follows the naming pattern:
\verb{\{engine\}_\{mode\}_\{parsnip_function\}()}
}
\seealso{
Other Internals:
\code{\link[=internal_make_fitted_wflw]{internal_make_fitted_wflw()}},
\code{\link[=internal_make_spec_tbl]{internal_make_spec_tbl()}},
\code{\link[=internal_make_wflw]{internal_make_wflw()}},
\code{\link[=internal_make_wflw_gee_lin_reg]{internal_make_wflw_gee_lin_reg()}},
\code{\link[=internal_make_wflw_predictions]{internal_make_wflw_predictions()}},
\code{\link{internal_model_builders_regression}},
\code{\link[=internal_set_args_to_tune]{internal_set_args_to_tune()}},
\code{\link[=make_classification_base_tbl]{make_classification_base_tbl()}},
\code{\link[=make_regression_base_tbl]{make_regression_base_tbl()}}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Internals}
