% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals-model-builders-regression.R
\name{internal_model_builders_regression}
\alias{internal_model_builders_regression}
\alias{lm_regression_linear_reg}
\alias{brulee_regression_linear_reg}
\alias{gee_regression_linear_reg}
\alias{glm_regression_linear_reg}
\alias{glmer_regression_linear_reg}
\alias{glmnet_regression_linear_reg}
\alias{gls_regression_linear_reg}
\alias{lme_regression_linear_reg}
\alias{lmer_regression_linear_reg}
\alias{stan_regression_linear_reg}
\alias{stan_glmer_regression_linear_reg}
\alias{cubist_regression_cubist_rules}
\alias{glm_regression_poisson_reg}
\alias{gee_regression_poisson_reg}
\alias{glmer_regression_poisson_reg}
\alias{glmnet_regression_poisson_reg}
\alias{hurdle_regression_poisson_reg}
\alias{stan_regression_poisson_reg}
\alias{stan_glmer_regression_poisson_reg}
\alias{zeroinfl_regression_poisson_reg}
\alias{earth_regression_bag_mars}
\alias{earth_regression_mars}
\alias{rpart_regression_bag_tree}
\alias{rpart_regression_decision_tree}
\alias{partykit_regression_decision_tree}
\alias{dbarts_regression_bart}
\alias{xgboost_regression_boost_tree}
\alias{lightgbm_regression_boost_tree}
\alias{mgcv_regression_gen_additive_mod}
\alias{nnet_regression_mlp}
\alias{brulee_regression_mlp}
\alias{kknn_regression_nearest_neighbor}
\alias{ranger_regression_rand_forest}
\alias{randomforest_regression_rand_forest}
\alias{xrf_regression_rule_fit}
\alias{liblinear_regression_svm_linear}
\alias{kernlab_regression_svm_linear}
\alias{kernlab_regression_svm_poly}
\alias{kernlab_regression_svm_rbf}
\title{Internal Model Builders for Regression}
\usage{
lm_regression_linear_reg()

brulee_regression_linear_reg()

gee_regression_linear_reg()

glm_regression_linear_reg()

glmer_regression_linear_reg()

glmnet_regression_linear_reg()

gls_regression_linear_reg()

lme_regression_linear_reg()

lmer_regression_linear_reg()

stan_regression_linear_reg()

stan_glmer_regression_linear_reg()

cubist_regression_cubist_rules()

glm_regression_poisson_reg()

gee_regression_poisson_reg()

glmer_regression_poisson_reg()

glmnet_regression_poisson_reg()

hurdle_regression_poisson_reg()

stan_regression_poisson_reg()

stan_glmer_regression_poisson_reg()

zeroinfl_regression_poisson_reg()

earth_regression_bag_mars()

earth_regression_mars()

rpart_regression_bag_tree()

rpart_regression_decision_tree()

partykit_regression_decision_tree()

dbarts_regression_bart()

xgboost_regression_boost_tree()

lightgbm_regression_boost_tree()

mgcv_regression_gen_additive_mod()

nnet_regression_mlp()

brulee_regression_mlp()

kknn_regression_nearest_neighbor()

ranger_regression_rand_forest()

randomforest_regression_rand_forest()

xrf_regression_rule_fit()

liblinear_regression_svm_linear()

kernlab_regression_svm_linear()

kernlab_regression_svm_poly()

kernlab_regression_svm_rbf()
}
\value{
A parsnip model specification object
}
\description{
Internal functions that build pre-configured parsnip model specifications
for regression models. Each function returns a parsnip model specification
with the mode and engine pre-set.
}
\details{
These functions are used internally by \code{internal_make_spec_tbl()} to create
model specifications. Each function follows the naming pattern:
\verb{\{engine\}_\{mode\}_\{parsnip_function\}()}
}
\seealso{
Other Internals:
\code{\link[=internal_make_fitted_wflw]{internal_make_fitted_wflw()}},
\code{\link[=internal_make_spec_tbl]{internal_make_spec_tbl()}},
\code{\link[=internal_make_wflw]{internal_make_wflw()}},
\code{\link[=internal_make_wflw_gee_lin_reg]{internal_make_wflw_gee_lin_reg()}},
\code{\link[=internal_make_wflw_predictions]{internal_make_wflw_predictions()}},
\code{\link{internal_model_builders_classification}},
\code{\link[=internal_set_args_to_tune]{internal_set_args_to_tune()}},
\code{\link[=make_classification_base_tbl]{make_classification_base_tbl()}},
\code{\link[=make_regression_base_tbl]{make_regression_base_tbl()}}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Internals}
