% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match-args.R
\name{match_args}
\alias{match_args}
\title{Match function arguments}
\usage{
match_args(f, args)
}
\arguments{
\item{f}{The parsnip function such as \code{"linear_reg"} as a string and without
the parentheses.}

\item{args}{The arguments you want to supply to \code{f}}
}
\value{
A list of matched arguments.
}
\description{
Match a functions arguments.
}
\details{
Match a functions arguments, the bad ones passed will be rejected but
the remaining passing ones will be returned.
}
\examples{

match_args(
  f = "linear_reg",
  args = list(
    mode = "regression",
    engine = "lm",
    trees = 1,
    mtry = 1
   )
 )

}
\seealso{
Other Utility:
\code{\link[=check_duplicate_rows]{check_duplicate_rows()}},
\code{\link[=core_packages]{core_packages()}},
\code{\link[=create_splits]{create_splits()}},
\code{\link[=create_workflow_set]{create_workflow_set()}},
\code{\link[=fast_classification_parsnip_spec_tbl]{fast_classification_parsnip_spec_tbl()}},
\code{\link[=fast_regression_parsnip_spec_tbl]{fast_regression_parsnip_spec_tbl()}},
\code{\link[=full_internal_make_wflw]{full_internal_make_wflw()}},
\code{\link[=install_deps]{install_deps()}},
\code{\link[=load_deps]{load_deps()}},
\code{\link[=quantile_normalize]{quantile_normalize()}}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Utility}
