% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-make-base-tbl.R
\name{make_regression_base_tbl}
\alias{make_regression_base_tbl}
\title{Utility Make Base Regression Tibble}
\usage{
make_regression_base_tbl()
}
\value{
A tibble
}
\description{
Creates a base tibble to create parsnip regression model specifications.
}
\details{
Creates a base tibble to create parsnip regression model specifications.
}
\examples{
make_regression_base_tbl()

}
\seealso{
Other Utility: 
\code{\link{core_packages}()},
\code{\link{create_splits}()},
\code{\link{fast_classification_parsnip_spec_tbl}()},
\code{\link{fast_regression_parsnip_spec_tbl}()},
\code{\link{install_deps}()},
\code{\link{load_deps}()},
\code{\link{make_classification_base_tbl}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Utility}
