% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-make-base-tbl.R
\name{make_regression_base_tbl}
\alias{make_regression_base_tbl}
\title{Internals Make Base Regression Tibble}
\usage{
make_regression_base_tbl()
}
\value{
A tibble
}
\description{
Creates a base tibble to create parsnip regression model specifications.
}
\details{
Creates a base tibble to create parsnip regression model specifications.
}
\examples{
make_regression_base_tbl()

}
\seealso{
Other Internals: 
\code{\link{internal_make_fitted_wflw}()},
\code{\link{internal_make_spec_tbl}()},
\code{\link{internal_make_wflw}()},
\code{\link{internal_make_wflw_gee_lin_reg}()},
\code{\link{internal_make_wflw_predictions}()},
\code{\link{internal_set_args_to_tune}()},
\code{\link{make_classification_base_tbl}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Internals}
